Revert mqttclient.h rename
