+++ esphome/components/mqtt/mqtt_client.cpp
@@ -1,4 +1,4 @@
-#include "mqttclient.h"
+#include "mqtt_client.h"
 
 #ifdef USE_MQTT
