+++ esphome/components/mqtt/mqtt_client.h
