+++ esphome/components/mqtt/mqtt_component.h
@@ -8,7 +8,7 @@
 
 #include "esphome/core/component.h"
 #include "esphome/core/entity_base.h"
-#include "mqttclient.h"
+#include "mqtt_client.h"
 
 namespace esphome {
 namespace mqtt {
