+++ esphome/components/mqtt_subscribe/sensor/mqtt_subscribe_sensor.h
@@ -6,7 +6,7 @@
 
 #include "esphome/core/component.h"
 #include "esphome/components/sensor/sensor.h"
-#include "esphome/components/mqtt/mqttclient.h"
+#include "esphome/components/mqtt/mqtt_client.h"
 
 namespace esphome {
 namespace mqtt_subscribe {
