+++ esphome/components/mqtt_subscribe/text_sensor/mqtt_subscribe_text_sensor.h
@@ -6,7 +6,7 @@
 
 #include "esphome/core/component.h"
 #include "esphome/components/text_sensor/text_sensor.h"
-#include "esphome/components/mqtt/mqttclient.h"
+#include "esphome/components/mqtt/mqtt_client.h"
 
 namespace esphome {
 namespace mqtt_subscribe {
