+++ esphome/components/status/status_binary_sensor.cpp
@@ -4,7 +4,7 @@
 #include "esphome/core/defines.h"
 
 #ifdef USE_MQTT
-#include "esphome/components/mqtt/mqttclient.h"
+#include "esphome/components/mqtt/mqtt_client.h"
 #endif
 #ifdef USE_API
 #include "esphome/components/api/api_server.h"
